requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
Django==3.2.5
django-cors-headers==3.7.0
<<<<<<< HEAD
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
=======
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
django-extensions==3.1.3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==3.2
>>>>>>> edbd58c ( - #10 Add(Signin): 로그인 EndPoint, API 구현)
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
<<<<<<< HEAD
jsonpointer==2.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
packaging==21.0
=======
packaging==21.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
openapi-codec==1.3.2
>>>>>>> edbd58c ( - #10 Add(Signin): 로그인 EndPoint, API 구현)
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pysha3==1.0.2
pytz==2021.1
PyYAML==5.4.1
<<<<<<< HEAD
requests==2.26.0
rfc3987==1.3.8
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
=======
rfc3987==1.3.8
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
requests==2.26.0
simplejson==3.17.3
>>>>>>> edbd58c ( - #10 Add(Signin): 로그인 EndPoint, API 구현)
six==1.16.0
sqlparse==0.4.1
strict-rfc3339==0.7
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
wcwidth==0.2.5
